Add solutions to challenge page
